// src/types/IAxieBriefs.ts
import {IAxieClass} from './IAxieClass';

export interface IAxieBriefs {
    BeastClass: IAxieClass[];
    AquaticClass: IAxieClass[];
    PlantClass: IAxieClass[];
    BirdClass: IAxieClass[];
    BugClass: IAxieClass[];
    ReptileClass: IAxieClass[];
    MechClass: IAxieClass[];
    DawnClass: IAxieClass[];
    DuskClass: IAxieClass[];
}